project/techtrends/requirements.txt
Flask==1.1.1
werkzeug==0.16.1
python-dotenv==0.18.0